% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-features.R
\name{daily_acf}
\alias{daily_acf}
\title{Features related to the autocorrelation lags for each day}
\usage{
daily_acf(x)
}
\arguments{
\item{x}{Time series of class \code{msts}.}
}
\value{
A list with the value of the autocorrelation for each day from 1 to
28.
}
\description{
Compute the autocorrelation lags for each day, from day 1 to 28.
}
