% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-metadata.R
\name{get_samples_per_day}
\alias{get_samples_per_day}
\title{List of samples per day per dataset}
\usage{
get_samples_per_day()
}
\value{
A list of samples per day per dataset.
}
\description{
Get a list of samples per day per dataset.
}
