% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-extraction.R
\name{get_timeseries_from_cooked_dataframe}
\alias{get_timeseries_from_cooked_dataframe}
\title{Time series from cooked dataframe}
\usage{
get_timeseries_from_cooked_dataframe(cdf)
}
\arguments{
\item{df}{Cooked dataframe.}
}
\value{
Multi-seasonal time series.
}
\description{
Creates a multi-seasonal time series \code{msts} from a coocked dataframe (or superior).
}
