% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-features.R
\name{load_factors}
\alias{load_factors}
\title{Features related to load factors}
\usage{
load_factors(x)
}
\arguments{
\item{x}{Time series of class \code{msts}.}
}
\value{
A list with the mean and variance of the load factors across seasonal periods.
}
\description{
Compute the load factor across seasonal periods (days, weeks and years). A high load factor means power usage is relatively constant. Low load factor shows that occasionally a high demand is set.
}
