% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-tools.R
\name{plot_kmeans}
\alias{plot_kmeans}
\title{Plotting of k-means results or elbow curve}
\usage{
plot_kmeans(km, feats_df = NULL, plot_clusters = FALSE, plot_elbow = FALSE)
}
\arguments{
\item{km}{An object of class `kmeans`.}

\item{feats_df}{Dataframe to be clustered (features, PCA scores).}

\item{plot_clusters}{If `TRUE`, plot k-means clusters.}

\item{plot_elbow}{If `TRUE`, plot elbow curve.}
}
\value{
Plotting of the k-means results or elbow curve.
}
\description{
Plot k-means results or elbow curve.
}
