% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-features.R
\name{stat_data_aggregates}
\alias{stat_data_aggregates}
\title{Features related to statistical data aggregates}
\usage{
stat_data_aggregates(x)
}
\arguments{
\item{x}{Time series of class \code{msts}.}
}
\value{
A list with statistical values of the different bins.
}
\description{
Compute means and standard deviations of time series aggregates, where bins relate to day, week and year periods.
}
\details{
\code{RE-WRITE THIS}
}
